Add in a control for optimization parameter
We added this in to control one of the parameters in the optimization models for constructing batch designs. This parameter, "quest_orth_log_coeff", is a parameter which goes with question orthogonality (aka question covariance) and is used to control question orthogonality in the objective function. When this parameter is negative, the optimization problem becomes unbounded. So we decide to make this parameter equal to 0 in the case where the user gives a negative value.