Update Batch_Design_and_Rollout
We updated the code for Batch_AO and Batch_MO in order to be able to handle a batch size of one. When first coded, Batch_AO and Batch_MO could handle batches of size one. However, a change was made involving the coefficient of the query covariance term which rendered having batch sizes of one impossible. This was noticed when rerunning the experiments and corrected.